+++ src/main/java/racingcar/model/domain/Car.java
@@ -1,4 +1,4 @@
-package racingcar.model.vo;
+package racingcar.model.domain;
 
 public class Car {
