+++ src/test/java/racingcar/model/domain/RacingTest.java
@@ -8,7 +8,6 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racingcar.model.vo.Car;
 
 public class RacingTest {
